Fold the transformer step into the ps-report-2 pupil data function to save a step
