--- func-ps-report/ps-report-3-transformer/function.json
@@ -1,19 +1,0 @@
-{
-  "bindings": [
-    {
-      "direction": "in",
-      "type": "serviceBusTrigger",
-      "name": "inputData",
-      "queueName": "ps-report-staging",
-      "connection": "AZURE_SERVICE_BUS_CONNECTION_STRING"
-    },
-    {
-      "direction": "out",
-      "type": "serviceBus",
-      "name": "outputData",
-      "queueName": "ps-report-export",
-      "connection": "AZURE_SERVICE_BUS_CONNECTION_STRING"
-    }
-  ],
-  "scriptFile": "../dist/functions-ps-report/ps-report-3-transformer/index.js"
-}
