+++ tslib/src/functions-ps-report/ps-report-2-pupil-data/mocks/check-config.ts
@@ -1,4 +1,4 @@
-import { type CheckConfig } from '../../../functions-ps-report/ps-report-2-pupil-data/models'
+import { type CheckConfig } from '../pupil-data.models'
 
 export const checkConfig: CheckConfig = {
   audibleSounds: false,
