+++ tslib/src/functions-ps-report/ps-report-2-pupil-data/mocks/check-incomplete.ts
@@ -1,4 +1,4 @@
-import { type Check } from '../../../functions-ps-report/ps-report-2-pupil-data/models'
+import { type Check } from '../pupil-data.models'
 
 export const check: Check = {
   id: 31,
