+++ tslib/src/functions-ps-report/ps-report-2-pupil-data/mocks/events.ts
@@ -1,4 +1,4 @@
-import type { Event } from '../../../functions-ps-report/ps-report-2-pupil-data/models'
+import type { Event } from '../pupil-data.models'
 import moment from 'moment'
 
 export const events: Event[] = [
