+++ tslib/src/functions-ps-report/ps-report-2-pupil-data/mocks/pupil-complete-and-restart-available-corrupt.ts
@@ -1,21 +1,21 @@
-import { type Pupil } from '../../../functions-ps-report/ps-report-2-pupil-data/models'
+import type { Pupil } from '../pupil-data.models'
 import moment from 'moment'
 
 export const pupil: Pupil = {
-  checkComplete: false,
-  currentCheckId: 1,
+  checkComplete: true,
+  currentCheckId: 6,
   dateOfBirth: moment.utc('2012-02-03'),
   forename: 'abc',
   gender: 'M',
   id: 3,
-  isEdited: false,
+  isEdited: true,
   jobId: 1,
   lastname: 'def',
-  middlenames: 'Other',
+  middlenames: 'Name',
   notTakingCheckReason: null,
   notTakingCheckCode: null,
-  restartAvailable: false,
+  restartAvailable: true,
   slug: 'abc-def-hij-klm',
   schoolId: 3,
-  upn: 'TEST985'
+  upn: 'TEST987'
 }
