+++ tslib/src/functions-ps-report/ps-report-2-pupil-data/mocks/pupil-not-attending-corrupt.ts
@@ -1,21 +1,21 @@
-import type { Pupil } from '../../ps-report-2-pupil-data/models'
+import type { Pupil } from '../pupil-data.models'
 import moment from 'moment'
 
 export const pupil: Pupil = {
-  checkComplete: null,
+  checkComplete: true,
   currentCheckId: null,
   dateOfBirth: moment.utc('2012-04-01'),
   forename: 'Tester',
   gender: 'M',
-  id: 3,
+  id: 4,
   isEdited: false,
   jobId: 1,
   lastname: 'Person',
-  notTakingCheckReason: 'Maladministration',
-  notTakingCheckCode: 'ANLLQ',
-  middlenames: 'middle name',
+  middlenames: 'Other',
+  notTakingCheckReason: 'Absent during check window',
+  notTakingCheckCode: 'ABSNT',
   restartAvailable: false,
   slug: 'abc-def-hij-klm',
-  schoolId: 3,
-  upn: 'TEST986'
+  schoolId: 4,
+  upn: 'TEST987'
 }
