+++ tslib/src/functions-ps-report/ps-report-2-pupil-data/mocks/pupil-who-completed-a-check.ts
@@ -1,4 +1,4 @@
-import { type Pupil } from '../../../functions-ps-report/ps-report-2-pupil-data/models'
+import { type Pupil } from '../pupil-data.models'
 import moment from 'moment'
 
 export const pupil: Pupil = {
