+++ tslib/src/functions-ps-report/ps-report-2-pupil-data/mocks/pupil-with-incomplete-check.ts
@@ -1,21 +1,21 @@
-import type { Pupil } from '../../../functions-ps-report/ps-report-2-pupil-data/models'
+import { type Pupil } from '../pupil-data.models'
 import moment from 'moment'
 
 export const pupil: Pupil = {
-  checkComplete: true,
-  currentCheckId: 6,
+  checkComplete: false,
+  currentCheckId: 1,
   dateOfBirth: moment.utc('2012-02-03'),
   forename: 'abc',
   gender: 'M',
   id: 3,
-  isEdited: true,
+  isEdited: false,
   jobId: 1,
   lastname: 'def',
-  middlenames: 'Name',
+  middlenames: 'Other',
   notTakingCheckReason: null,
   notTakingCheckCode: null,
-  restartAvailable: true,
+  restartAvailable: false,
   slug: 'abc-def-hij-klm',
   schoolId: 3,
-  upn: 'TEST987'
+  upn: 'TEST985'
 }
