+++ tslib/src/functions-ps-report/ps-report-2-pupil-data/ps-report.service.spec.ts
@@ -9,7 +9,7 @@ describe.skip('PsReportService', () => {
   let logger: ILogger
   let psReportDataService: IPsReportDataService
   const schoolUuid = 'AAAA-BBBB-CCCC-DDDD'
-  const outputBindings: IOutputBinding = { psReportPupilMessage: [] }
+  const outputBindings: IOutputBinding = { psReportExportOutput: [] }
   const mockPupils = [{ id: 1, schoolId: 99 }, { id: 2, schoolId: 99 }, { id: 3, schoolId: 99 }]
   const mockSchool = { id: 99, name: 'test school' }
   const psReportSchoolFanOutMessage: PsReportSchoolFanOutMessage = {
@@ -27,7 +27,7 @@ describe.skip('PsReportService', () => {
       getPupils: jest.fn(),
       getSchool: jest.fn()
     }
-    outputBindings.psReportPupilMessage = []
+    outputBindings.psReportExportOutput = []
 
     sut = new PsReportService(outputBindings, logger, psReportDataService)
   })
@@ -70,6 +70,6 @@ describe.skip('PsReportService', () => {
       .mockResolvedValueOnce({ data: 2 })
       .mockResolvedValueOnce({ data: 3 })
     await sut.process(psReportSchoolFanOutMessage)
-    expect(outputBindings.psReportPupilMessage).toHaveLength(3)
+    expect(outputBindings.psReportExportOutput).toHaveLength(3)
   })
 })
