+++ tslib/src/functions-ps-report/ps-report-2-pupil-data/ps-report.service.ts
@@ -1,9 +1,10 @@
 import { type IPsReportDataService, PsReportDataService } from './ps-report.data.service'
-import { type Pupil, type PupilResult, type School } from './models'
+import { type Pupil, type PupilResult, type School } from './pupil-data.models'
 import { type ILogger } from '../../common/logger'
 import { type IOutputBinding } from '.'
 import type { PsReportSchoolFanOutMessage } from '../common/ps-report-service-bus-messages'
 import config from '../../config'
+import { ReportLine } from './report-line.class'
 
 const logName = 'ps-report-2-pupil-data: PsReportService'
 
@@ -45,7 +46,7 @@ export class PsReportService {
       }
       try {
         const result: PupilResult = await this.dataService.getPupilData(pupil, school)
-        const output: PupilResult = {
+        const pupilResult: PupilResult = {
           answers: result.answers,
           check: result.check,
           checkConfig: result.checkConfig,
@@ -55,7 +56,13 @@ export class PsReportService {
           pupil: result.pupil,
           school: result.school
         }
-        this.outputBinding.psReportPupilMessage.push(output)
+
+        // Now we have the pupil data we can transform it into the report format
+        const reportLine = new ReportLine(pupilResult.answers, pupilResult.check, pupilResult.checkConfig, pupilResult.checkForm, pupilResult.device, pupilResult.events, pupilResult.pupil, pupilResult.school)
+        const outputData = reportLine.transform()
+
+        // Send the transformed pupil data onto the ps-report-export queue using the output bindings.
+        this.outputBinding.psReportExportOutput.push(outputData)
       } catch (error: any) {
         // Ignore the error on the particular pupil and carry on so it reports on the rest of the school
         this.logger.error(`${logName}: ERROR: Failed to retrieve pupil data for pupil ${pupil.slug} in school ${incomingMessage.uuid}
