+++ tslib/src/functions-ps-report/ps-report-3b-stage-csv-file/index.ts
@@ -3,7 +3,7 @@ import { performance } from 'perf_hooks'
 import * as sb from '@azure/service-bus'
 import config from '../../config'
 import * as RA from 'ramda-adjunct'
-import { type IPsychometricReportLine } from '../ps-report-3-transformer/models'
+import { type IPsychometricReportLine } from '../ps-report-2-pupil-data/transformer-models'
 import { jsonReviver } from '../../common/json-reviver'
 import { PsReportStagingDataService } from './ps-report-staging.data.service'
 import { CsvTransformer } from './csv-transformer'
